+++ src/test/java/com/exmachina/sam/currency/scheduledtasks/helpers/RatesMapperTest.java
@@ -5,6 +5,7 @@ import com.exmachina.sam.currency.scheduledtasks.pojos.ThirdPartyRates;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -26,22 +27,22 @@ public class RatesMapperTest {
                         rate -> rate.getDestination(),
                         rate -> rate.getCoefficient())
                 .containsExactlyInAnyOrder(
-                        tuple("source_currency_1" , "destination_currency_1" , 3.14),
-                        tuple("source_currency_1" , "destination_currency_2" , 1.59),
-                        tuple("source_currency_1" , "destination_currency_3" , 2.65),
-                        tuple("source_currency_2" , "destination_currency_1" , 3.14),
-                        tuple("source_currency_2" , "destination_currency_2" , 1.59),
-                        tuple("source_currency_2" , "destination_currency_3" , 2.65)
+                        tuple("source_currency_1" , "destination_currency_1" , BigDecimal.valueOf(3.14)),
+                        tuple("source_currency_1" , "destination_currency_2" , BigDecimal.valueOf(1.59)),
+                        tuple("source_currency_1" , "destination_currency_3" , BigDecimal.valueOf(2.65)),
+                        tuple("source_currency_2" , "destination_currency_1" , BigDecimal.valueOf(3.14)),
+                        tuple("source_currency_2" , "destination_currency_2" , BigDecimal.valueOf(1.59)),
+                        tuple("source_currency_2" , "destination_currency_3" , BigDecimal.valueOf(2.65))
                 );
     }
 
     private ThirdPartyRates createThirdPartyRatesWithFakes(){
 
-        LinkedHashMap<String, Double> destinationsAndCoefficients = new LinkedHashMap<String, Double>()
+        LinkedHashMap<String, BigDecimal> destinationsAndCoefficients = new LinkedHashMap<String, BigDecimal>()
         {{
-            put("destination_currency_1", 3.14);
-            put("destination_currency_2", 1.59);
-            put("destination_currency_3", 2.65);
+            put("destination_currency_1", BigDecimal.valueOf(3.14));
+            put("destination_currency_2", BigDecimal.valueOf(1.59));
+            put("destination_currency_3", BigDecimal.valueOf(2.65));
         }};
 
         return new ThirdPartyRates(){{
